requirements.txt
# requirements.txt

requests
mysql-connector-python
python-dotenv
phonenumbers